+++ vite.config.ts
@@ -19,7 +19,7 @@ export default defineConfig({
           build: {
             outDir: 'out/main',
             rollupOptions: {
-              external: ['better-sqlite3', 'electron', 'electron-updater'],
+              external: ['node-sqlite3-wasm', 'electron', 'electron-updater'],
             },
           },
         },